// backend/src/types/order.types.ts
export type DeliveryStatus =
  | 'REQUESTED'
  | 'ACCEPTED'
  | 'READY_FOR_PICKUP'
  | 'PICKED_UP'
  | 'IN_TRANSIT'
  | 'DELIVERED'
  | 'CANCELLED'
// ...
export interface DeliveryInput {
  fullName: string
  phone: string
  city: string
  subCity: string
  neighborhood?: string
  deliveryLocation: string
  deliveryNotes?: string
}